% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_border.R
\name{get_vals_with_border}
\alias{get_vals_with_border}
\title{Add some border to the matrix of values
No idea if that is really working....}
\usage{
get_vals_with_border(star, add.x, add.y)
}
\description{
Add some border to the matrix of values
No idea if that is really working....
}
