% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDEM.R
\name{prepareDEM}
\alias{prepareDEM}
\title{Function to prepare DEM for Blender}
\usage{
prepareDEM(
  demSource = NULL,
  bbSource = NULL,
  resX = NULL,
  resY = NULL,
  add.x = NULL,
  add.y = NULL
)
}
\description{
Function to prepare DEM for Blender
}
